pyproject.toml
[project]
name = "gg-translate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi>=0.115.14",
    "jose>=1.0.0",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "sentencepiece>=0.2.0",
    "sqlalchemy>=2.0.41",
    "torch>=2.7.1",
    "transformers>=4.53.1",
    "uvicorn>=0.35.0",
]
